refactor: extract some part of temporal logic to interlocutor ICIJ/datashare#2134
