+++ datashare-tasks/src/main/java/org/icij/datashare/asynctasks/TaskManagerTemporal.java
@@ -1,24 +1,18 @@
 package org.icij.datashare.asynctasks;
 
-import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
-import io.temporal.api.enums.v1.WorkflowIdConflictPolicy;
-import io.temporal.api.enums.v1.WorkflowIdReusePolicy;
-import io.temporal.client.*;
+import io.temporal.client.WorkflowExecutionAlreadyStarted;
 import io.temporal.common.SearchAttributes;
-import org.icij.datashare.asynctasks.temporal.TemporalInputPayload;
 import org.icij.datashare.asynctasks.temporal.TemporalInterlocutor;
 import org.icij.datashare.tasks.RoutingStrategy;
 
 import java.io.IOException;
 import java.io.Serializable;
-import java.time.Duration;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static io.temporal.api.enums.v1.WorkflowExecutionStatus.WORKFLOW_EXECUTION_STATUS_RUNNING;
 import static org.icij.datashare.LambdaExceptionUtils.rethrowConsumer;
 import static org.icij.datashare.asynctasks.Task.State.FINAL_STATES;
 import static org.icij.datashare.asynctasks.temporal.TemporalInterlocutor.*;
@@ -29,11 +23,8 @@ public class TaskManagerTemporal implements TaskManager {
     private final RoutingStrategy routingStrategy;
     private final TaskRepository taskRepository;
 
-    private static final Duration DEFAULT_WORKFLOW_TASK_TIMEOUT = Duration.ofDays(7);
-
     // TODO: add support for continue-as-new https://docs.temporal.io/develop/java/continue-as-new
 
-    private static final String TERMINATION_MSG = "terminated_by_user";
     public static final String WORKFLOWS_DEFAULT = "default-java";
 
     public TaskManagerTemporal(TemporalInterlocutor temporal, TaskRepository taskRepository, RoutingStrategy routingStrategy) {
@@ -45,27 +36,18 @@ public class TaskManagerTemporal implements TaskManager {
     @Override
     public <V extends Serializable> String startTask(Task<V> taskView, Group group) throws TaskAlreadyExists {
         String taskId = taskView.id;
-        WorkflowOptions.Builder optionBuilder = WorkflowOptions.newBuilder().setWorkflowId(taskId)
-            .setWorkflowTaskTimeout(DEFAULT_WORKFLOW_TASK_TIMEOUT) // TODO: set this per task
-            .setWorkflowIdReusePolicy(WorkflowIdReusePolicy.WORKFLOW_ID_REUSE_POLICY_REJECT_DUPLICATE)
-            .setWorkflowIdConflictPolicy(WorkflowIdConflictPolicy.WORKFLOW_ID_CONFLICT_POLICY_FAIL)
-            .setTypedSearchAttributes(generateSearchAttributes(taskView))
-            .setTaskQueue(resolveWfTaskQueue(taskView.name, group));
         try {
-            temporal.newUntypedWorkflowStub(taskView.name, optionBuilder.build())
-                .start(new TemporalInputPayload(taskView.args));
+            temporal.createWorkflow(taskId, taskView.name, resolveWfTaskQueue(taskView.name ,group),
+                    generateSearchAttributes(taskView), taskView.args);
         } catch (WorkflowExecutionAlreadyStarted ex) {
             throw new TaskAlreadyExists(taskId, ex);
         }
-
-        // Super important force description to refresh the cache and make the task visible
-        temporal.newUntypedWorkflowStub(taskId).describe();
         return taskId;
     }
 
     @Override
     public <V extends Serializable> Task<V> getTask(String taskId) throws IOException, UnknownTask {
-        return temporal.parseTask(getWorkflowExecution(taskId), this);
+        return temporal.parseTask(temporal.getWorkflowExecution(taskId), this);
     }
 
     @Override
@@ -93,19 +75,7 @@ public class TaskManagerTemporal implements TaskManager {
 
     @Override
     public boolean stopTask(String taskId) throws IOException, UnknownTask {
-        // TODO: add support for cancellation rather than termination, update the TaskManager API accordingly
-        return unknownIfNotFound(tId -> {
-            WorkflowStub workflowStub = temporal.newUntypedWorkflowStub(tId);
-            boolean isRunning = workflowStub.describe().getStatus() == WORKFLOW_EXECUTION_STATUS_RUNNING;
-            try {
-                workflowStub.terminate(TERMINATION_MSG);
-            } catch (WorkflowNotFoundException ex) {
-                if (!ex.getCause().getMessage().contains("workflow execution already completed")) {
-                    throw ex;
-                }
-            }
-            return isRunning;
-        }, taskId);
+        return temporal.terminateWorkflow(taskId);
     }
 
     @Override
@@ -170,12 +140,6 @@ public class TaskManagerTemporal implements TaskManager {
         return resolveWfTaskQueue(routingStrategy, taskName, group);
     }
 
-    private WorkflowExecutionDescription getWorkflowExecution(String taskId) throws UnknownTask {
-        return unknownIfNotFound(t -> {
-            return temporal.newUntypedWorkflowStub(taskId).describe();
-        }, taskId);
-    }
-
     static SearchAttributes generateSearchAttributes(Task<?> taskView) {
         SearchAttributes.Builder builder = SearchAttributes.newBuilder()
             .set(PROGRESS_CUSTOM_ATTRIBUTE, 0d)
@@ -204,27 +168,4 @@ public class TaskManagerTemporal implements TaskManager {
         throw new RuntimeException("failed to clear task in " + timeout + " " + timeUnit);
     }
 
-    protected void awaitExecutionDeletion(Set<String> taskIds) throws InterruptedException {
-        // TODO: avoid the infinite loop
-        // TODO: implement throttling
-        while (true) {
-            boolean allDeleted = taskIds.stream().allMatch(tId -> {
-                try {
-                    temporal.newUntypedWorkflowStub(tId).describe();
-                } catch (StatusRuntimeException ex) {
-                    if (ex.getStatus().getCode().equals(Status.Code.NOT_FOUND)) {
-                        return true;
-                    }
-                } catch (WorkflowNotFoundException ignored) {
-                    return true;
-                }
-                return false;
-            });
-            if (allDeleted) {
-                break;
-            }
-            // TODO: define a proper duration
-            Thread.sleep(DEFAULT_NAMESPACE_POLL_INTERVAL.toMillis());
-        }
-    }
 }
